+++ src/Module/Api/Method/Api3/Albums3Method.php
@@ -27,7 +27,6 @@ namespace Ampache\Module\Api\Method\Api3;
 
 use Ampache\Module\Api\Xml3_Data;
 use Ampache\Module\Api\Api;
-use Ampache\Module\Api\Api3;
 use Ampache\Module\System\Session;
 use Ampache\Repository\Model\User;
 
@@ -45,7 +44,7 @@ final class Albums3Method
      */
     public static function albums(array $input)
     {
-        $browse = Api3::getBrowse();
+        $browse = Api::getBrowse();
         $browse->reset_filters();
         $browse->set_type('album');
         $browse->set_sort('name', 'ASC');