+++ src/Module/Api/Method/Api3/Songs3Method.php
@@ -26,7 +26,6 @@ declare(strict_types=0);
 namespace Ampache\Module\Api\Method\Api3;
 
 use Ampache\Module\Api\Api;
-use Ampache\Module\Api\Api3;
 use Ampache\Module\Api\Xml3_Data;
 use Ampache\Module\System\Session;
 use Ampache\Repository\Model\User;
@@ -45,7 +44,7 @@ final class Songs3Method
      */
     public static function songs(array $input)
     {
-        $browse = Api3::getBrowse();
+        $browse = Api::getBrowse();
         $browse->reset_filters();
         $browse->set_type('song');
         $browse->set_sort('title', 'ASC');