+++ src/Module/Api/Method/Api4/Genres4Method.php
@@ -26,7 +26,6 @@ declare(strict_types=0);
 namespace Ampache\Module\Api\Method\Api4;
 
 use Ampache\Module\Api\Api;
-use Ampache\Module\Api\Api4;
 use Ampache\Module\Api\Json4_Data;
 use Ampache\Module\Api\Xml4_Data;
 
@@ -51,7 +50,7 @@ final class Genres4Method
      */
     public static function genres(array $input)
     {
-        $browse = Api4::getBrowse();
+        $browse = Api::getBrowse();
         $browse->reset_filters();
         $browse->set_type('tag');
         $browse->set_sort('name', 'ASC');