+++ src/Module/Api/Method/Api4/PlaylistCreate4Method.php
@@ -47,7 +47,7 @@ final class PlaylistCreate4Method
      * This create a new playlist and return it
      *
      * @param array $input
-     * name = (string) Alpha-numeric search term
+     * name = (string) Playlist name
      * type = (string) 'public'|'private'
      * @return boolean
      */