+++ src/Module/Api/Method/Api4/Podcasts4Method.php
@@ -60,7 +60,7 @@ final class Podcasts4Method
 
             return false;
         }
-        $browse = Api4::getBrowse();
+        $browse = Api::getBrowse();
         $browse->reset_filters();
         $browse->set_type('podcast');
         $browse->set_sort('title', 'ASC');