+++ src/Module/Api/Method/PlaylistCreateMethod.php
@@ -48,7 +48,7 @@ final class PlaylistCreateMethod
      * This create a new playlist and return it
      *
      * @param array $input
-     * name = (string) Alpha-numeric search term
+     * name = (string) Playlist name
      * type = (string) 'public', 'private'
      * @return boolean
      */